+++ sync_camera_disk/main.py
@@ -20,7 +20,7 @@ from sync_camera_disk.config import Config, Destination, Source, SourceType, Syn
 from . import source
 from .destination import DatedFolderDestination
 from .filter_disks import filter_disks_to_syncs
-from .operation import perform_operation
+from .operation import OperationType, perform_operation
 
 app = typer.Typer()
 
@@ -119,9 +119,10 @@ def show_syncs(
 @app.command()
 def sync(
     config_path: Annotated[
-        Path, typer.Argument(help="Path to probes.yml config")
+        Path, typer.Argument(help="Path to config.yml config")
     ] = DEFAULT_CONFIG_PATH,
     dry_run: bool = True,
+    log_identical_operations: bool = True,
 ) -> None:
     """Sync files from disks to configured destinations"""
     config = parse_yaml_file_as(Config, config_path)
@@ -151,12 +152,16 @@ def sync(
                 for operation in destination.generate_operations(file_set=file_set):
                     counters[str(operation.operation)] += 1
                     LOG.debug("operation", operation=operation)
-                    LOG.info(
-                        operation.operation,
-                        type=sync.source.type,
-                        source=operation.source,
-                        destination=operation.destination,
-                    )
+                    if (
+                        operation.operation == OperationType.identical
+                        and log_identical_operations
+                    ) or (operation.operation != OperationType.identical):
+                        LOG.info(
+                            operation.operation,
+                            type=sync.source.type,
+                            source=operation.source,
+                            destination=operation.destination,
+                        )
                     result = perform_operation(operation, dry_run=dry_run)
                     LOG.debug("operation result", result=result, success=result.success)
                     if not result.success: