+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix handling of sub folders when copying files (in particular the ATEM ISO structure). Was flattening the folder structures.
 
 ## [0.7.0] - 2024-09-18
 ### Added